Zfs support

+++ main.go
@@ -5,7 +5,11 @@ import (
 	"log"
 	"os"
 	"os/exec"
+	"path"
 )
+
+const Zfs = "zfs"
+const Ext4 = "ext4"
 
 func crash(err error) {
 	if err != nil {
@@ -24,6 +28,7 @@ func sh(cmdName string, args ...string) {
 
 func main() {
 
+	rootFileSystem := flag.String("fs", Ext4, "filesystem to use on root, currently ext4 and zfs")
 	targetDevice := flag.String("device", "", "Device to use ")
 	flag.Parse()
 
@@ -36,10 +41,18 @@ func main() {
 	sh("parted", *targetDevice, "--", "mkpart", "primary", "linux-swap", "-8GiB", "100%")
 	sh("parted", *targetDevice, "--", "mkpart", "ESP", "fat32", "1MiB", "512MiB")
 	sh("parted", *targetDevice, "--", "set", "3", "esp", "on")
-	sh("mkfs.ext4", rootPartition)
-	sh("mkswap", swapPartition)
-	sh("mkfs.fat", "-F", "32", "-n", "boot", bootPartition)
-	sh("mount", rootPartition, "/mnt")
+	if *rootFileSystem == Ext4 {
+		sh("mkfs.ext4", rootPartition)
+		sh("mkswap", swapPartition)
+		sh("mkfs.fat", "-F", "32", "-n", "boot", bootPartition)
+		sh("mount", rootPartition, "/mnt")
+	} else if *rootFileSystem == Zfs {
+		zfsPoolName := "zroot"
+		nixosZfsDataset := path.Join(zfsPoolName, "root")
+		sh("zpool", "create", "-O", "mountpoint=none", "-O", "atime=off", "-O", "compression=lz4", "-O", "xattr=sa", "-O", "acltype=posixacl", "-o", "ashift=12", "-R", "/mnt", zfsPoolName, rootPartition)
+		sh("zfs", "create", "-o", "mountpoint=none", nixosZfsDataset)
+		sh("mount", "-t", "zfs", nixosZfsDataset, "/mnt")
+	}
 	sh("mkdir", "-p", "/mnt/boot")
 	sh("mount", bootPartition, "/mnt/boot")
 	sh("swapon", swapPartition)
